% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shades.R
\name{create_shades}
\alias{create_shades}
\title{Create a set of shades from a palette}
\usage{
create_shades(palette, n = 10)
}
\arguments{
\item{palette}{a vector of hex colours that make up a base palette}

\item{n}{the number of shades to calculate}
}
\description{
This function takes an vector of hex colours and creates a set of shades
for each colour. The function is still experimental and may look to change
the internals of how the colours are determined.
}
