% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\docType{data}
\name{default_pal}
\alias{default_pal}
\title{Default Palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
default_pal
}
\description{
Default Palette
}
\keyword{datasets}
