% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolors.R
\name{get_coolors}
\alias{get_coolors}
\title{Get Coolors}
\usage{
get_coolors(url)
}
\arguments{
\item{url}{the url for a specific coolors palette}
}
\value{
a vector of hex colours
}
\description{
Get a vector of hex colours from a coolors url
}
