% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shades.R
\name{get_shades}
\alias{get_shades}
\title{Get a set of shades from a base colour}
\usage{
get_shades(colour, n = 10)
}
\arguments{
\item{colour}{a hex colour value to determine a set of shades from.}

\item{n}{the number of shades to calculate.}
}
\value{
a vector of hex colours
}
\description{
Get a set of shades from a base colour
}
