% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shades.R
\name{hex2hsl}
\alias{hex2hsl}
\title{Convert Hex Colours to HSL}
\usage{
hex2hsl(colour)
}
\arguments{
\item{colour}{a vector of hex colours in the format '#123456'}
}
\value{
a list object vectors representing the HSL values of the colours
}
\description{
Convert Hex Colours to HSL
}
