% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shades.R
\name{hsl2rgb}
\alias{hsl2rgb}
\title{Convert HSL values to RGB}
\usage{
hsl2rgb(hsl)
}
\arguments{
\item{hsl}{a vector of length three with values for Hue, Saturation & Light}
}
\description{
Convert HSL values to RGB
}
