% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palette.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot a set of charts to visualise a palette in the wild}
\usage{
plot_palette(palette = default_pal, theme = theme_plot(), seed = 2022)
}
\arguments{
\item{palette}{a vector of hex colours for plotting}

\item{theme}{a ggplot theme to assist with the look and feel}

\item{seed}{a seed to keep the plots consistent}
}
\description{
\code{plot_palette()} is designed to streamline the process of finding palettes
that will work well with real data. It uses \code{ggplot2} to create a set of
charts to demonstrate the look and feel of a specific palette
}
