% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolors.R
\name{print_coolors}
\alias{print_coolors}
\title{Print a Coolors Palette}
\usage{
print_coolors(url)
}
\arguments{
\item{url}{a palette url from the coolors.co website}
}
\description{
This function shows the colours in the palette with their associated HEX,
RGB and HSL values
}
