% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_palette.R
\name{print_palette}
\alias{print_palette}
\title{Print a Colour Palette}
\usage{
print_palette(palette = default_pal)
}
\arguments{
\item{palette}{a vector of hex colours for plotting}
}
\description{
This function shows the colours in the palette with their associated HEX,
RGB and HSL values
}
