% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shades.R
\name{rgb2hsl}
\alias{rgb2hsl}
\title{Convert an RGB colour to an HSL value}
\usage{
rgb2hsl(col)
}
\arguments{
\item{col}{a vector of length three with values for Red, Green & Blue. The RGB assumes a range from 0-255.}
}
\description{
Convert an RGB colour to an HSL value
}
