% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_qld}
\alias{sf_qld}
\title{Queensland Shapefule}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 80 rows and 6 columns.
}
\source{
Australian Bureau of Statistics - https://www.abs.gov.au/AUSSTATS/abs@.nsf/Lookup/1270.0.55.003Main+Features1June\%202020?OpenDocument
}
\usage{
sf_qld
}
\value{
sf object
}
\description{
Queensland Shapefule
}
\keyword{datasets}
