% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{theme_plot}
\alias{theme_plot}
\title{Default Theme for plotcolr charts}
\usage{
theme_plot(
  base_family = "Arial",
  base_size = 14,
  plot_title_family = "Arial Bold",
  plot_title_size = 28,
  subtitle_family = base_family,
  subtitle_size = 16,
  title_family = base_family,
  title_size = 14,
  caption_family = base_family,
  caption_size = 12
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_size}{plot title font family and size}

\item{subtitle_family, subtitle_size}{subtitle font family and size}

\item{title_family, title_size}{subtitle font family and size}

\item{caption_family, caption_size}{caption font family and size}
}
\description{
A default ggplot2 theme to help with the presentation of colour palettes
}
