% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolors.R
\name{plot_coolors}
\alias{plot_coolors}
\title{Plot Colours from Coolors website}
\usage{
plot_coolors(
  url,
  theme = theme_plot(),
  seed = 2022,
  cvd = c("none", "deutan", "protan", "tritan", "all")
)
}
\arguments{
\item{url}{a palette url from the coolors.co website}

\item{theme}{a ggplot theme to assist with the look and feel}

\item{seed}{a seed to keep the plots consistent}

\item{cvd}{a colour vision deficiency to simulate}
}
\value{
a ggplot object
}
\description{
Visualise a palette from a coolors url.
}
