% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palette.R
\name{save_plots}
\alias{save_plots}
\title{Save Palette Plots}
\usage{
save_plots(filename, plot = last_plot())
}
\arguments{
\item{filename}{the name and location of the file to be saved}

\item{plot}{Plot to save, defaults to last plot displayed.}
}
\description{
Save plots of palettes with sensible defaults
}
